+++ packages/cater-runtime/index.js
@@ -1,8 +1,6 @@
 // Copyright Jon Williams 2017-2018. See LICENSE file.
-const BabelOptions = require('./src/options-babel.js');
-const DefaultOptions = require('./src/options-default.js');
-const fs = require('fs');
-const path = require('path');
+const Config = require('./src/config');
+const DefaultConfig = require('./src/config-default.js');
 const HandlerCater = require('./src/handler-cater');
 const HandlerLogging = require('./src/handler-logging');
 const HttpServer = require('./src/http-server');
@@ -16,28 +14,15 @@ const RuntimeCater = require('./src');
  * @see module/cater.
  */
 
-function loadConfig(file = null) {
-  const configFile = file || path.join(process.cwd(), 'cater.config.js');
-  if (!fs.existsSync(configFile)) return {};
-  // eslint-disable-next-line global-require, import/no-dynamic-require
-  const options = require(configFile);
+function index(providedConfig = null) {
+  const config = Config(providedConfig, DefaultConfig());
 
-  if (options.env) {
-    const env = options.env[process.env.NODE_ENV];
-    if (env) Object.assign(options, env);
+  if (config.babelRegisterDisable) {
+    const babelConfig = config.babel;
+    babelConfig.ignore = /\/node_modules\//;
+    require('babel-register')(babelConfig); // eslint-disable-line global-require
   }
-
-  return options;
-}
-
-function index(options = null) {
-  const caterOptions = options || loadConfig();
-  if (caterOptions.disableBabel) {
-    const babelOptions = BabelOptions();
-    babelOptions.ignore = /\/node_modules\//;
-    require('babel-register')(babelOptions); // eslint-disable-line global-require
-  }
-  return new RuntimeCater(caterOptions);
+  return new RuntimeCater(config);
 }
 
 function readyCommandLine() {
@@ -46,12 +31,11 @@ function readyCommandLine() {
 }
 
 const exporting = {
-  BabelOptions,
-  DefaultOptions,
+  Config,
+  DefaultConfig,
   HandlerCater,
   HandlerLogging,
   HttpServer,
-  loadConfig,
   Middleware,
   readyCommandLine,
   RuntimeCater