+++ packages/cater-runtime/test/fixtures/cater.config.simple.js
@@ -1,3 +1,3 @@
 module.exports = {
-  httpPort: 8080
+  httpPort: 3333
 };